moved things around to avoid cyclical imports
