+++ sssom/parsers.py
@@ -11,8 +11,9 @@ import yaml
 from rdflib import Graph
 
 from .sssom_document import MappingSet, Mapping, MappingSetDocument
-from .util import get_file_extension, RDF_FORMATS
-from io import StringIO
+from .datamodel_util import get_file_extension
+from .util import RDF_FORMATS
+from sssom.datamodel_util import read_pandas
 
 cwd = os.path.abspath(os.path.dirname(__file__))
 
@@ -289,36 +290,7 @@ def get_parsing_function(input_format, filename):
     else:
         raise Exception(f'Unknown input format: {input_format}')
 
-def read_csv(filename, comment='#', sep=','):
-    lines = "".join([line for line in open(filename) 
-                     if not line.startswith(comment)])
-    return pd.read_csv(StringIO(lines), sep=sep)
 
-def read_pandas(filename: str, sep=None) -> pd.DataFrame:
-    """
-    wrapper to pd.read_csv that handles comment lines correctly
-    :param filename:
-    :param sep: File separator in pandas (\t or ,)
-    :return:
-    """
-    if not sep:
-        extension = get_file_extension(filename)
-        sep = "\t"
-        if extension == "tsv":
-            sep = "\t"
-        elif extension == "csv":
-            sep = ","
-        else:
-            logging.warning(f"Cannot automatically determine table format, trying tsv.")
-
-    # from tempfile import NamedTemporaryFile
-    # with NamedTemporaryFile("r+") as tmp:
-    #    with open(filename, "r") as f:
-    #        for line in f:
-    #            if not line.startswith('#'):
-    #                tmp.write(line + "\n")
-    #    tmp.seek(0)
-    return read_csv(filename, comment='#', sep=sep).fillna("")
 
 
 def _prepare_mapping(mapping: Mapping):
