+++ sssom/util.py
@@ -2,8 +2,7 @@ import pandas as pd
 import random
 import hashlib
 from sssom.sssom_datamodel import Entity, Mapping
-from sssom.datamodel_util import MappingSetDiff, EntityPair
-#from sssom.parsers import read_pandas
+from sssom.datamodel_util import read_pandas, MappingSetDiff, EntityPair
 from typing import Dict, Tuple, List
 
 # TODO: use sssom_datamodel
@@ -194,14 +193,6 @@ def dataframe_to_ptable(df: pd.DataFrame, priors=[0.02, 0.02, 0.02, 0.02], inver
     return rows
 
 RDF_FORMATS=['ttl', 'turtle', 'nt']
-
-def get_file_extension(filename: str) -> str:
-    parts = filename.split(".")
-    if len(parts) > 0:
-        f_format = parts[-1]
-        return f_format
-    else:
-        raise Exception(f'Cannot guess format from {filename}')
 
 def sha256sum(filename):
     h  = hashlib.sha256()
