+++ tests/test_conversion.py
@@ -6,10 +6,11 @@ import pandas as pd
 
 from rdflib import Graph
 from sssom.sssom_document import MappingSetDocument
-from sssom.parsers import read_pandas, get_parsing_function
+from sssom.parsers import get_parsing_function
 from sssom.writers import to_owl_graph, to_rdf_graph, to_dataframe, to_jsonld_dict
 from sssom.writers import write_json, write_rdf, write_owl, write_tsv
 from .test_data import ensure_test_dir_exists, SSSOMTestCase, get_all_test_cases
+from sssom.datamodel_util import read_pandas
 
 
 class SSSOMReadWriteTestSuite(unittest.TestCase):
